ml_project/requirements.txt
antlr4-python3-runtime==4.8
astroid==2.5.6
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
hydra-core==1.0.6
importlib-resources==5.1.2
iniconfig==1.1.1
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
numpy==1.20.2
omegaconf==2.0.6
packaging==20.9
pandas==1.2.4
Pillow==8.2.0
pluggy==0.13.1
py==1.10.0
pylint==2.8.2
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
toml==0.10.2
torch==1.8.1
torchvision==0.9.1
tqdm==4.60.0
typing-extensions==3.7.4.3
wrapt==1.12.1
